requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
APScheduler==3.10.4
redis==5.0.1
requests==2.31.0
aiohttp==3.9.1
